+++ contracts/src/Steel.sol
@@ -159,7 +159,7 @@ library Encoding {
 library ChainSpec {
     uint256 internal constant ETHEREUM_MAINNET_CHAIN_ID = 1;
     uint256 internal constant ETHEREUM_SEPOLIA_CHAIN_ID = 11155111;
-    uint256 internal constant ETHEREUM_HOLESKY_CHAIN_ID = 17000;
+    uint256 internal constant ETHEREUM_HOODI_CHAIN_ID = 560048;
     uint256 internal constant STEEL_TEST_PRAGUE_CHAIN_ID = 5733100018;
 
     /// @dev Error selector: 0x45b21e77
@@ -172,13 +172,13 @@ library ChainSpec {
     // TODO(povw): Add something to keep this in sync with the Rust.
     function configID(uint256 chainID) internal pure returns (bytes32) {
         if (chainID == ETHEREUM_MAINNET_CHAIN_ID) {
-            return hex"9a223c7ca04c969f1cacbe5b8db44c308b2c53390505d3d48c834ed4469fc839";
+            return hex"47dc59f84afd2e9e7a48c4012004ab7c77fbd9acf822bf1143b8442c6c8851d4";
         }
         if (chainID == ETHEREUM_SEPOLIA_CHAIN_ID) {
-            return hex"5c9552dc9bfad8572ded4f818bb35b0f4260660c1554236986b768ae999b4b60";
+            return hex"90c1e882b1f0fda4dc7f1c66c07ed3d2a74e443834905faa9f32f583b71f459d";
         }
-        if (chainID == ETHEREUM_HOLESKY_CHAIN_ID) {
-            return hex"8eae1ba5f877e6ad7007bf6985f5245be7d758457fb4eb7e6a72d47f49bea389";
+        if (chainID == ETHEREUM_HOODI_CHAIN_ID) {
+            return hex"34cb1defd939572b00439d2c13f93c033b82227067371c910ad104d527c78860";
         }
         if (chainID == STEEL_TEST_PRAGUE_CHAIN_ID) {
             return hex"33e32d9590cd4b168773ca27de65d535f2e744274b1437acb712dd4264f2eb87";